Move iframe unwrapper to extractor
